requirements.txt
chardet==5.2.0
pandas==2.2.3
PyYAML==6.0.2
scikit-learn==1.6.0
torch==2.5.1